backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.1
cymem==2.0.11
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
langcodes==3.5.0
language_data==1.3.0
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.13
nltk==3.9.1
numpy==2.3.3
openai==1.109.1
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
preshed==3.0.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pypdfium2==4.30.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
shellingham==1.5.4
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.48.0
thinc==8.3.6
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.3
